backend/requirements.txt
fastapi
uvicorn
SQLAlchem
email-validator
psycopg2-binary